Move InventoryIncrementalTasksRunner to common package
